split unpatch tests out of test-index.js
